requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pypdf
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
python-multipart
sqlalchemy
python-dotenv
tiktoken
jinja2
google-generativeai
google-cloud-texttospeech